requirements.txt
transformer
peft
gradio
tqdm
sklearn
pandas
numpy
